[REFACTOR]: estrutura de pastas
