+++ components/global/Footer/Copyright/index.tsx
