+++ components/global/Footer/Shared/index.tsx
@@ -1,5 +1,5 @@
 import Link from 'next/link';
-import icons from '../../../lib/icons';
+import icons from '../../../../lib/icons';
 
 const Shared = () => {
   return (
