+++ components/global/Footer/index.tsx
