+++ components/global/Hero.tsx
