+++ components/layout.tsx
@@ -1,0 +1,16 @@
+import Footer from './global/Footer';
+import NavBar from './global/NavBar';
+
+interface LayoutProps {
+  children: JSX.Element;
+}
+
+export default function Layout({ children }: LayoutProps) {
+  return (
+    <>
+      <NavBar />
+      <main>{children}</main>
+      <Footer />
+    </>
+  );
+}
