--- containers/layout.tsx
@@ -1,12 +1,0 @@
-import Footer from 'components/Footer';
-import NavBar from 'components/NavBar';
-
-export default function Layout({ children }) {
-  return (
-    <>
-      <NavBar />
-      <main>{children}</main>
-      <Footer />
-    </>
-  );
-}
