+++ pages/_app.tsx
@@ -1,4 +1,4 @@
-import Layout from 'containers/layout';
+import Layout from 'components/layout';
 import type { AppProps } from 'next/app';
 import '../styles/globals.css';
